+++ Packages/com.unity.inputsystem/InputSystem/Plugins/OnScreen/OnScreenButton.cs
@@ -1,4 +1,5 @@
 #if PACKAGE_DOCS_GENERATION || UNITY_INPUT_SYSTEM_ENABLE_UI
+using System;
 using UnityEngine.EventSystems;
 using UnityEngine.InputSystem.Layouts;
 
@@ -22,6 +23,16 @@ namespace UnityEngine.InputSystem.OnScreen
         public void OnPointerDown(PointerEventData eventData)
         {
             SendValueToControl(1.0f);
+            m_Pressed = true;
+        }
+
+        protected void OnDisable()
+        {
+            if (m_Pressed)
+            {
+                SendValueToControl(0f);
+                m_Pressed = false;
+            }
         }
 
         ////TODO: pressure support
@@ -38,6 +49,8 @@ namespace UnityEngine.InputSystem.OnScreen
         [InputControl(layout = "Button")]
         [SerializeField]
         private string m_ControlPath;
+
+        [NonSerialized] bool m_Pressed;
 
         protected override string controlPathInternal
         {